+++ src/apis/tracker/tracksim.py
@@ -888,6 +888,8 @@ async def post_update(response: Response, request: Request):
                 for item in truck_damage.keys():
                     damage += nfloat(truck_damage[item])
 
+                damage = damage * app.config.economy.wear_ratio
+
                 await app.db.execute(dhrid, f"UPDATE economy_truck SET odometer = odometer + {driven_distance}, damage = damage + {damage}, income = income + {economy_revenue} WHERE vehicleid = {vehicleid}")
                 if current_damage + damage > app.config.economy.max_wear_before_service:
                     await app.db.execute(dhrid, f"UPDATE economy_truck SET status = -1 WHERE vehicleid = {vehicleid}")